reduce count sql when get page with not full content

+++ src/main/java/com/github/longdt/vertxorm/repository/PageRequest.java
@@ -48,7 +48,7 @@ public class PageRequest {
     }
 
     @JsonIgnore
-    public int getOffset() {
+    public long getOffset() {
         return (index - 1) * size;
     }
 }
